+++ src/main/java/nextstep/favorite/application/RestTemplateConfig.java
@@ -1,0 +1,13 @@
+package nextstep.favorite.application;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class RestTemplateConfig {
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+}
