+++ src/test/java/nextstep/utils/GithubTestController.java
@@ -3,12 +3,12 @@ package nextstep.utils;
 import nextstep.favorite.application.dto.GithubAccessTokenRequest;
 import nextstep.favorite.application.dto.GithubAccessTokenResponse;
 import nextstep.favorite.application.dto.GithubProfileResponse;
-import nextstep.favorite.application.dto.GithubResponses;
+import nextstep.member.acceptance.GithubResponses;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-public class TestController {
+public class GithubTestController {
     @PostMapping("/github/login/oauth/access_token")
     public ResponseEntity<GithubAccessTokenResponse> accessToken(@RequestBody GithubAccessTokenRequest tokenRequest) {
         GithubResponses githubUser = GithubResponses.findByCode(tokenRequest.getCode());
